+++ Sources/VCore/Views and ViewControllers/Infinite Scrolling UI Views/Infinite Scrolling UI Collection View/InfiniteScrollingUICollectionViewActivityIndicatorViewUIModel.swift
@@ -1,0 +1,15 @@
+//
+//  InfiniteScrollingUICollectionViewActivityIndicatorViewUIModel.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 11/21/21.
+//
+
+#if canImport(UIKit) && !os(watchOS)
+
+import Foundation
+
+// MARK: - Infinite Scrolling UI Collection View Activity Indicator View UI Model
+typealias InfiniteScrollingUICollectionViewActivityIndicatorViewUIModel = InfiniteScrollingUITableViewActivityIndicatorViewUIModel
+
+#endif