+++ Sources/VCore/Views and ViewControllers/Infinite Scrolling UI Views/Infinite Scrolling UI Collection View/InfiniteScrollingUICollectionViewPaginationState.swift
@@ -1,5 +1,5 @@
 //
-//  InfiniteScrollingTableViewPaginationState.swift
+//  InfiniteScrollingCollectionViewPaginationState.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 11/21/21.
@@ -7,10 +7,10 @@
 
 #if canImport(UIKit) && !os(watchOS)
 
-import Foundation
+import UIKit
 
-// MARK: - Infinite Scrolling Table View Pagination State
-extension InfiniteScrollingTableView {
+// MARK: - Infinite Scrolling UI Collection View Pagination State
+extension InfiniteScrollingUICollectionView {
     public typealias PaginationState = InfiniteScrollingPaginationState
 }